Refactor and reorganize utility functions and imports
- Moved `setupEnv` and `asEnv` functions to a new `container.ts` file for better organization of container-related utilities.
- Updated imports across various files to reflect the new structure, ensuring consistency and clarity.
- Removed the deprecated `env.ts` and `index.ts` files to streamline the codebase.
- Introduced a new `auth.ts` file for handling token refresh logic, improving modularity in authentication utilities.
