+++ __tests__/lib/utils/setupEnv.test.ts
@@ -8,7 +8,7 @@ jest.mock("node:child_process", () => {
 
 import { exec as execCallback } from "node:child_process"
 
-import { setupEnv } from "@/lib/utils/setupEnv"
+import { setupEnv } from "@/lib/utils/container"
 
 // Cast for convenience – the mock is created above via jest.mock
 const execMock = execCallback as unknown as jest.MockedFunction<
