+++ lib/utils/auth.ts
