+++ lib/utils/client.ts
@@ -1,0 +1,2 @@
+// Client-side utilities - frontend/browser-specific helpers
+export { cn, getApiKeyFromLocalStorage } from "./utils-common"
