+++ lib/utils/utils-common.ts
@@ -1,3 +1,5 @@
+// TODO: Migrate to /lib/utils/client.ts or /lib/utils/server.ts
+
 import { type ClassValue, clsx } from "clsx"
 import { EventEmitter } from "events"
 import { twMerge } from "tailwind-merge"
