+++ lib/workflows/identifyPRGoal.ts
@@ -12,7 +12,7 @@ import {
   createWorkflowStateEvent,
 } from "@/lib/neo4j/services/event"
 import { initializeWorkflowRun } from "@/lib/neo4j/services/workflow"
-import { createGetIssueTool } from "@/lib/tools"
+import { createGetIssueTool } from "@/lib/tools/GetIssueTool"
 import { GitHubIssue } from "@/lib/types/github"
 
 interface IdentifyPRGoalParams {
